pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "try-on-anything"
version = "1.1.0"
description = "Try On Anything - AI-powered virtual try-on system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["accessory_try_on*"]


[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
